+++ .eslintrc.js
@@ -73,6 +73,7 @@ const tsConfigs = {
 
     '@typescript-eslint/no-redundant-type-constituents': 'off',
     '@typescript-eslint/no-unsafe-enum-comparison': 'off',
+    '@typescript-eslint/no-this-alias': 'off',
     '@typescript-eslint/consistent-type-imports': 'error',
     '@typescript-eslint/member-delimiter-style': [
       2,