+++ packages/orchestrator/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+
+- `microfrontend-loader` webcomponent which wraps the `micro-lc` loading API to reuse MFE loading inside compose applications
+- extraction of the application configuration maker which is now shared with `microfrontend-loader`
+
 ### Fixes
 
 - security updates / vulnerability fixes in `.docker/Dockerfile` related with `nginx:1.25.1-alpine` for `CVE-2023-2975`, `CVE-2023-3138`, and `CVE-2023-3316`